pyproject.toml
[tool.poetry]
name = "dedither"
version = "0.1.0"
description = ""
authors = ["freetz13 <alexander.astafiev.1989@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow-SIMD = "^9.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
